i do the bonus

+++ DAL/Customer.cs
@@ -11,8 +11,8 @@ namespace IDAL
             public int ID { get; set; }
             public string name { get; set; }
             public string phone { get; set; }
-            public double longitude { get; set; }
-            public double lattitude { get; set; }
+            public float longitude { get; set; }
+            public float lattitude { get; set; }
 
             public override string ToString()
             {
@@ -20,7 +20,6 @@ namespace IDAL
                 result += $"ID is: {ID}, \n";
                 result += $"Name is: {name}, \n";
                 result += $"Telephone is: {phone.Substring(0, 3) + '-' + phone.Substring(3)}, \n";
-               // result += $"Telephone is: {phone}, \n";
                 result += $"Lattitude is: {lattitude}, \n";
                 result += $"longitude is: {longitude}, \n";
                 return result;
